Switches away from Prettierx (#1792)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Prettierx hasn't had a release since 2021, and the latest prettier has
no issues with how the code is at the moment. I'm just going to make an
executive decision and move us to the latest prettier version as I'm
truthfully unaware of the circumstances that caused TG to use prettierx
instead of prettier main.

## Why It's Good For The Game
Allows devs to actually format the TGUI files in a way that passes
ESLint

## Changelog

:cl:
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->