Fixes Kepori and Vox prosthetics when spawning (#1769)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Fixes: https://github.com/shiptest-ss13/Shiptest/issues/1426

Apparently, checking for a species before it is set doesn't work,
imagine if we had known this 6 months ago...
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
allows people to use prosthetics on their Kepori and Vox instead of
spawning with human limbs
## Changelog

:cl:
fix: fixed Kepori and Vox prosthetics when spawning
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Mark Suckerberg <29362068+MarkSuckerberg@users.noreply.github.com>
Co-authored-by: Mark Suckerberg <29362068+MarkSuckerberg@users.noreply.github.com>