Adds an extra check to prevent admins from spawning planets on top of others (#1816)
## About The Pull Request

Lets admins choose an overmap position or force a random position when
spawning planets/ruins.

## Why It's Good For The Game

Potentially fixes #1746 
Giving admins the ability to choose a location for their planets and
ruins is probably cool.

## Changelog

:cl:
fix: Potentially fixed an issue with planets swapping type
admin: Added the option for admins to choose the location of new planets
/:cl: