room underneath the captain's room is empty
the room underneath the captain's room is now a janitorial closet, so moved the mop from the engi office to the closet. also adds  a wallsafe with intel, moves scrubbers and wiring around so it makes More Sense, gives the crew a sleepy time box, some other small stuff that made me angry and i forgot to list here. also supervisor lizrd