Remaps the Osprey (#1827)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![nanotrasen_osprey2](https://user-images.githubusercontent.com/60533805/225514436-600eb517-9e9f-4de6-898d-e1bc49b9a0e4.png)

![nanotrasen_osprey2_areas](https://user-images.githubusercontent.com/60533805/225514442-4c2b0cb9-d046-495a-a850-5425124833f7.png)

A full remap of the Osprey, featuring an improved layout and some slight
improvements to starting equipment.
The changes are as follows:
- Disposals has been upgraded to a functional deliveries system
- Dedicated AI core
- Improved cargo & sci layout
- Better-equipped kitchen with a central location
- Far larger engineering and atmos spaces with less wasted space in the
back of the ship
- A prebuilt telecomms relay tuned to Nanotrasen channels
- Ship now flies the right way

## Why It's Good For The Game
The Osprey was looking pretty sad, aged poorly, and had grown quite
unpopular. This should hopefully help fix the issue.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
tweak: Fully remapped the Osprey
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Apogee-dev <60533805+Apogee-dev@users.noreply.github.com>
Co-authored-by: Bjarl <94164348+Bjarl@users.noreply.github.com>