Gecko Quality of Life (#1842)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Gives the Gecko some much-needed quality of life adjustments.

- Cargo bays now have holofield projectors
- airlock now has a dedicated air canister to refill faster
- burn chamber no longer has air alarm inside chamber
- fulltire firelocks replaced with pairs of directional firelocks

## Why It's Good For The Game

the gecko has a strong concept and good aesthetics, but fell flat on
actually being a reasonably functional ship. Standardizing the cargo
bays and smoothing out some of the rougher functionality issues should
make it much less of a headache to play.

## Changelog

:cl:
tweak: Gecko-class Salvage Runner now has holofields and an improved
airlock
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->