Datumises Planet Types & Adds Ruin Weighting Framework (#1803)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Does what it says on the tin. Moves the planet types out of a massive
switch statement, and into a datum for each type. Also puts all ruin
types into a singular list of lists with the type as the key instead of
adding a new var to the mapping subsystem and hooking it up like that.

Also adds a system for ruin selection weighting, everything is an even 1
for right now though so it's no real gameplay change yet.

## Why It's Good For The Game
Allows for even easier creation of ruin types and planet types, which I
suppose is needed.

## Changelog

:cl:
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Mark Suckerberg <29362068+MarkSuckerberg@users.noreply.github.com>