Jupiter-Class Stormrider (#1788)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a new pirate ship, the Jupiter-Class Stormrider.
It's designed as an AI centric raiding ship, with the ability to dive
into storms to recharge its powerbanks.
![2023 03 28-22 37
02](https://user-images.githubusercontent.com/94164348/228412101-48d75444-a7f9-4c41-a3eb-ad5b4303b327.png)
![2023 03 28-22 37
07](https://user-images.githubusercontent.com/94164348/228412110-2f589ea2-ce7c-47a8-8d1b-d30157b08ea6.png)


https://user-images.githubusercontent.com/94164348/221093940-d3ce5262-8fe0-4877-bafb-e04381c32b44.mp4
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Pirate.... good?
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: An unknown warship has been spotted lurking in electrical storms.
Be cautious, starfarers.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: spockye <79304582+spockye@users.noreply.github.com>